refactor: move footer and navbar into their own sub-directories
