+++ client/components/Navbar/Navbar.spec.js
