+++ client/components/index.js
@@ -3,8 +3,8 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
-export {default as Navbar} from './Layout/Navbar'
-export {default as Footer} from './Layout/Footer'
+export {default as Navbar} from './Navbar/Navbar'
+export {default as Footer} from './Footer/Footer'
 export {Login, Signup} from './Auth/AuthForm'
 export {default as MealList} from './Meal/MealList'
 export {default as Notifier, openSnackbar} from './Utils/Notifier'
